Refactor url base classes:
fix #221, fix #222, close #330
