+++ doc/qbk/quickref.xml
@@ -72,6 +72,7 @@
           <member><link linkend="url.ref.boost__urls__segments_view">segments_view</link></member>
           <member><link linkend="url.ref.boost__urls__static_url">static_url</link></member>
           <member><link linkend="url.ref.boost__urls__url">url</link></member>
+          <member><link linkend="url.ref.boost__urls__url_base">url_base</link></member>
           <member><link linkend="url.ref.boost__urls__url_view">url_view</link></member>
         </simplelist>
