+++ example/route/route.cpp
@@ -11,6 +11,7 @@
 #include <boost/url/segments_encoded.hpp>
 #include <boost/url/segments_encoded_view.hpp>
 #include <boost/url/string_view.hpp>
+#include <boost/url/url.hpp>
 #include <boost/url/url_view.hpp>
 #include <boost/url/static_url.hpp>
 #include <boost/filesystem/fstream.hpp>
