+++ include/boost/url/impl/static_url.ipp
@@ -10,6 +10,7 @@
 #ifndef BOOST_URL_IMPL_STATIC_URL_IPP
 #define BOOST_URL_IMPL_STATIC_URL_IPP
 
+#include <boost/url/static_url.hpp>
 #include <boost/url/url_view.hpp>
 #include <boost/url/detail/except.hpp>
 #include <boost/assert.hpp>
@@ -18,73 +19,43 @@ namespace boost {
 namespace urls {
 
 static_url_base::
-~static_url_base()
+static_url_base(
+    char* buf,
+    std::size_t cap) noexcept
 {
-    // prevent url from
-    // deallocating memory
-    s_ = nullptr;
+    s_ = buf;
+    cap_ = cap;
+    s_[0] = '\0';
+    u_.cs_ = s_;
 }
 
 static_url_base::
 static_url_base(
     char* buf,
-    std::size_t cap) noexcept
-    : url(buf, cap)
+    std::size_t cap,
+    string_view s)
+    : static_url_base(buf, cap)
 {
+    copy(parse_uri_reference(s).value());
 }
 
 void
 static_url_base::
-copy(url const& u)
+clear_impl() noexcept
 {
-    this->url::copy(u);
+    u_ = detail::url_impl(false);
+    s_[0] = '\0';
+    u_.cs_ = s_;
 }
 
 void
 static_url_base::
-construct(string_view s)
+reserve_impl(
+    std::size_t n)
 {
-    auto r = parse_uri_reference(s);
-    if(r.has_error())
-        detail::throw_invalid_argument(
-            BOOST_CURRENT_LOCATION);
-    this->url::operator=(r.value());
-}
-
-void
-static_url_base::
-copy(url_view const& u)
-{
-    this->url::copy(u);
-}
-
-url_view
-static_url_base::
-convert() const noexcept
-{
-    return url_view(*this);
-}
-
-char*
-static_url_base::
-allocate(std::size_t n)
-{
-    (void)n;
-    // should never get here
-    BOOST_ASSERT(
-        n > capacity_in_bytes());
-    detail::throw_invalid_argument(
-        "alloc_impl",
-        BOOST_CURRENT_LOCATION);
-}
-
-void
-static_url_base::
-deallocate(char*)
-{
-    // should never get here
-    detail::throw_invalid_argument(
-        "free_impl",
+    if(n <= cap_)
+        return;
+    detail::throw_bad_alloc(
         BOOST_CURRENT_LOCATION);
 }
