+++ include/boost/url/url.hpp
@@ -12,33 +12,12 @@
 #define BOOST_URL_URL_HPP
 
 #include <boost/url/detail/config.hpp>
-#include <boost/url/ipv4_address.hpp>
-#include <boost/url/ipv6_address.hpp>
-#include <boost/url/params.hpp>
-#include <boost/url/params_encoded.hpp>
-#include <boost/url/scheme.hpp>
-#include <boost/url/segments.hpp>
-#include <boost/url/segments_encoded.hpp>
-#include <boost/url/url_view.hpp>
-#include <boost/url/detail/any_path_iter.hpp>
-#include <boost/url/detail/any_query_iter.hpp>
-#include <cstdint>
-#include <iosfwd>
-#include <memory>
-#include <string>
+#include <boost/url/url_base.hpp>
+#include <boost/assert.hpp>
 #include <utility>
 
 namespace boost {
 namespace urls {
-
-#ifndef BOOST_URL_DOCS
-namespace detail {
-struct any_path_iter;
-}
-namespace grammar {
-class lut_chars;
-}
-#endif
 
 /** A modifiable container for a URL.
 
@@ -80,37 +59,11 @@ class lut_chars;
         @ref resolve.
 */
 class BOOST_SYMBOL_VISIBLE url
-    : public url_view
+    : public url_base
 {
-    friend class urls::segments;
-    friend class urls::params;
-    friend class segments_encoded;
-    friend class params_encoded;
+    friend std::hash<url>;
 
-#ifndef BOOST_URL_DOCS
-protected:
-#endif
-    char* s_ = nullptr;
-    std::size_t cap_ = 0;
-
-    BOOST_URL_DECL
-    url(char* buf,
-        std::size_t cap) noexcept;
-
-    BOOST_URL_DECL
-    void
-    copy(url_view const& u);
-
-    BOOST_URL_DECL
-    virtual
-    char*
-    allocate(
-        std::size_t new_cap);
-
-    BOOST_URL_DECL
-    virtual
-    void
-    deallocate(char* s);
+    using url_view_base::digest;
 
 public:
     //--------------------------------------------
@@ -126,7 +79,6 @@ public:
         invalidated.
     */
     BOOST_URL_DECL
-    virtual
     ~url();
 
     /** Constructor
@@ -152,6 +104,32 @@ public:
     BOOST_URL_DECL
     url() noexcept;
 
+    /** Construct from a string
+
+        This function constructs a URL from
+        the string `s`, which must contain a
+        valid URI or <em>relative-ref</em> or
+        else an exception is thrown.
+
+        @par BNF
+        @code
+        URI           = scheme ":" hier-part [ "?" query ] [ "#" fragment ]
+
+        relative-ref  = relative-part [ "?" query ] [ "#" fragment ]
+        @endcode
+
+        @throw std::invalid_argument parse error.
+
+        @param s The string to parse.
+
+        @par Specification
+        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-4.1"
+            >4.1. URI Reference</a>
+    */
+    BOOST_URL_DECL
+    explicit
+    url(string_view s);
+
     /** Constructor
 
         This function performs a move-construction
@@ -173,7 +151,10 @@ public:
         @param u The url to construct from.
     */
     BOOST_URL_DECL
-    url(url const& u);
+    url(url_view_base const& u)
+    {
+        copy(u);
+    }
 
     /** Constructor
 
@@ -181,8 +162,11 @@ public:
 
         @param u The url to construct from.
     */
-    BOOST_URL_DECL
-    url(url_view const& u);
+    url(url const& u)
+        : url(static_cast<
+            url_view_base const&>(u))
+    {
+    }
 
     /** Assignment
 
@@ -211,9 +195,10 @@ public:
         @param u The url to copy.
     */
     url&
-    operator=(url const& u)
+    operator=(
+        url_view_base const& u)
     {
-        *this = u.base();
+        copy(u);
         return *this;
     }
 
@@ -228,1437 +213,19 @@ public:
 
         @param u The url to copy.
     */
-    BOOST_URL_DECL
     url&
-    operator=(url_view const& u);
-
-    /** Construct from a string
-
-        This function constructs a URL from
-        the string `s`, which must contain a
-        valid URI or <em>relative-ref</em> or
-        else an exception is thrown.
-
-        @par BNF
-        @code
-        URI           = scheme ":" hier-part [ "?" query ] [ "#" fragment ]
-
-        relative-ref  = relative-part [ "?" query ] [ "#" fragment ]
-        @endcode
-
-        @throw std::invalid_argument parse error.
-
-        @param s The string to parse.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-4.1"
-            >4.1. URI Reference</a>
-    */
-    BOOST_URL_DECL
-    url(string_view s);
-
-    //--------------------------------------------
-
-    /** Return the encoded URL as a null-terminated string
-    */
-    char const*
-    c_str() const noexcept
+    operator=(url const& u)
     {
-        return u_.cs_;
-    }
-
-    /** Returns the total number of bytes currently available to the container
-    */
-    std::size_t
-    capacity_in_bytes() const noexcept
-    {
-        return cap_;
-    }
-
-    /** Adjust the capacity without changing the size
-
-        This function adjusts the capacity
-        of the container in bytes, without
-        affecting the current contents.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @throw bad_alloc Allocation failure
-
-        @par bytes The number of bytes to reserve.
-    */
-    BOOST_URL_DECL
-    void
-    reserve_bytes(
-        std::size_t bytes);
-
-    /** Clear the contents.
-    
-        @par Exception Safety
-        No-throw guarantee.
-    */
-    BOOST_URL_DECL
-    void
-    clear() noexcept;
-
-    //--------------------------------------------
-    //
-    // Scheme
-    //
-    //--------------------------------------------
-
-private:
-    void set_scheme_impl(string_view s, urls::scheme id);
-public:
-
-    /** Remove the scheme
-
-        This function removes the scheme if it
-        is present.
-
-        @par BNF
-        @code
-        URI           = scheme ":" hier-part [ "?" query ] [ "#" fragment ]
-        @endcode
-
-        @par Exception Safety
-        Throws nothing.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.1">
-            3.1. Scheme (rfc3986)</a>
-
-        @see @ref set_scheme.
-    */
-    BOOST_URL_DECL
-    url&
-    remove_scheme() noexcept;
-
-    /** Set the scheme
-
-        This function sets the scheme to the specified
-        string, which must contain a valid scheme
-        without a trailing colon (':'). If the scheme
-        is invalid, an exception is thrown.
-
-        @par Example
-        @code
-        url u = parse_uri( "http://www.example.com" );
-
-        u.set_scheme( "https" );         // u is now "https://www.example.com"
-
-        assert( u.string() == "https://www.example.com" );
-
-        u.set_scheme( "1forall");       // throws, invalid scheme
-        @endcode
-
-        @par BNF
-        @code
-        scheme        = ALPHA *( ALPHA / DIGIT / "+" / "-" / "." )
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The scheme to set.
-
-        @throw std::invalid_argument invalid scheme.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.1">
-            3.1. Scheme (rfc3986)</a>
-
-        @see
-            @ref remove_scheme.
-    */
-    BOOST_URL_DECL
-    url&
-    set_scheme(string_view s);
-
-    /** Set the scheme
-
-        This function sets the scheme to the specified
-        known @ref urls::scheme id, which may not be
-        @ref scheme::unknown or else an exception is
-        thrown. If the id is @ref scheme::none, this
-        function behaves as if @ref remove_scheme
-        were called.
-
-        @par Example
-        @code
-        url u;
-        u.set_scheme( scheme::http );           // produces "http:"
-        u.set_scheme( scheme::none );           // produces ""
-        u.set_scheme( scheme::unknown);         // throws, invalid scheme
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param id The scheme to set.
-
-        @throw std::invalid_argument invalid scheme.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.1">
-            3.1. Scheme (rfc3986)</a>
-    */
-    BOOST_URL_DECL
-    url&
-    set_scheme(urls::scheme id);
-
-    //--------------------------------------------
-    //
-    // Authority
-    //
-    //--------------------------------------------
-
-private:
-    char* set_user_impl(std::size_t n);
-public:
-
-    /** Set the user.
-
-        The user is set to the specified string,
-        replacing any previous user. If a userinfo
-        was not present it is added, even if the
-        user string is empty. The resulting URL
-        will have an authority if it did not have
-        one previously.
-
-        Any special or reserved characters in the
-        string are automatically percent-encoded.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set. This string may
-        contain any characters, including nulls.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.1">
-            3.2.1. User Information (rfc3986)</a>
-    */
-    BOOST_URL_DECL
-    url&
-    set_user(
-        string_view s);
-
-    /// @copydoc set_user()
-    BOOST_URL_DECL
-    url&
-    set_user(
-        pct_encoded_view s);
-
-    /** Set the user.
-
-        The user is set to the specified string,
-        replacing any previous user. If a userinfo
-        was not present it is added, even if the
-        user string is empty. The resulting URL
-        will have an authority if it did not have
-        one previously.
-
-        The string must be a valid percent-encoded
-        string for the user field, otherwise an
-        exception is thrown.
-
-        @par BNF
-        @code
-        user          = *( unreserved / pct-encoded / sub-delims )
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set.
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_user(
-        string_view s);
-
-private:
-    char* set_password_impl(std::size_t n);
-public:
-
-    /** Remove the password
-    */
-    BOOST_URL_DECL
-    url&
-    remove_password() noexcept;
-
-    /** Set the password.
-
-        This function sets the password to the specified
-        string, replacing any previous password:
-
-        @li If the string is empty, the password is
-        cleared, and the first occurring colon (':') is
-        removed from the userinfo if present, otherwise
-
-        @li If ths string is not empty then the password
-        is set to the new string.
-        Any special or reserved characters in the
-        string are automatically percent-encoded.
-        If the URL previously did not have an authority
-        (@ref has_authority returns `false`), a double
-        slash ("//") is prepended to the userinfo.
-
-        @par Exception Safety
-
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set. This string may
-        contain any characters, including nulls.
-    */
-    BOOST_URL_DECL
-    url&
-    set_password(
-        string_view s);
-
-    /// @copydoc set_password()
-    BOOST_URL_DECL
-    url&
-    set_password(
-        pct_encoded_view s);
-
-    /** Set the password.
-
-        The password is set to the encoded string `s`,
-        replacing any previous password:
-
-        @li If the string is empty, the password is
-        cleared, and the first occurring colon (':') is
-        removed from the userinfo if present, otherwise
-
-        @li If ths string is not empty then the password
-        is set to the new string.
-        If the URL previously did not have an authority
-        (@ref has_authority returns `false`), a double
-        slash ("//") is prepended to the userinfo.
-        The string must meet the syntactic requirements
-        of <em>password</em> otherwise an exception is
-        thrown.
-
-        @par ANBF
-        @code
-        password      = *( unreserved / pct-encoded / sub-delims / ":" )
-        @endcode
-
-        @par Exception Safety
-
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set.
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_password(
-        string_view s);
-
-private:
-    char* set_userinfo_impl(std::size_t n);
-public:
-
-    /** Remove the userinfo
-    */
-    BOOST_URL_DECL
-    url&
-    remove_userinfo() noexcept;
-
-    /** Set the userinfo.
-
-        Sets the userinfo of the URL to the given
-        encoded string:
-
-        @li If the string is empty, the userinfo is
-        cleared, else
-
-        @li If the string is not empty, then the userinfo
-        is set to the given string. The user is set to
-        the characters up to the first colon if any,
-        while the password is set to the remaining
-        characters if any.
-        If the URL previously did not have an authority
-        (@ref has_authority returns `false`), a double
-        slash ("//") is prepended to the userinfo.
-        The string must meet the syntactic requirements
-        of <em>userinfo</em> otherwise an exception is
-        thrown.
-
-        @par BNF
-        @code
-        userinfo      = [ [ user ] [ ':' password ] ]
-
-        user          = *( unreserved / pct-encoded / sub-delims )
-        password      = *( unreserved / pct-encoded / sub-delims / ":" )
-        @endcode
-
-        @par Exception Safety
-
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set.
-
-        @throw std::exception Thrown on error
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_userinfo(
-        string_view s);
-
-    /** Set the userinfo.
-
-        The userinfo is set to the specified string,
-        replacing any previous userinfo. If a userinfo
-        was not present it is added, even if the
-        userinfo string is empty. The resulting URL
-        will have an authority if it did not have
-        one previously.
-
-        Any special or reserved characters in the
-        string are automatically percent-encoded.
-
-        The interpretation of userinfo as
-        individual user and password components
-        is scheme-dependent. Transmitting
-        passwords in URLs is deprecated.
-
-        This function treats userinfo as a unit
-        independent of the user and password
-        components.
-
-        If the input string has a ":" character,
-        its first occurrence will be considered
-        the separator between the user and the
-        password fields.
-
-        If the username might contain a ":", the
-        functions @ref set_user and
-        @ref set_password should be used
-        separately so that the character can be
-        unambiguously encoded.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set. This string may
-        contain any characters, including nulls.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.1">
-            3.2.1. User Information (rfc3986)</a>
-    */
-    BOOST_URL_DECL
-    url&
-    set_userinfo(
-        string_view s);
-
-    /// @copydoc set_userinfo()
-    BOOST_URL_DECL
-    url&
-    set_userinfo(
-        pct_encoded_view s);
-
-    //--------------------------------------------
-
-private:
-    char* set_host_impl(std::size_t n);
-public:
-
-    /** Set the host
-
-        The host is set to the specified IPv4,
-        address, replacing any previous host. If
-        an authority was not present, it is added.
-
-        @par Postconditions
-        @code
-        this->host_type() == host_type::ipv4 && this->ipv4_address() == addr
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.2">
-            3.2.2. Host (rfc3986)</a>
-
-        @param addr The address to set.
-    */
-#ifdef BOOST_URL_DOCS
-    url& set_host( ipv4_address const& addr );
-#else
-    BOOST_URL_DECL
-    url&
-    set_host(
-        urls::ipv4_address const& addr);
-#endif
-
-    /** Set the host
-
-        The host is set to the specified IPv6,
-        address, replacing any previous host.
-        If an authority did not
-        previously exist it is added by prepending
-        a double slash ("//") at the beginning of
-        the URL or after the scheme if a scheme is
-        present.
-
-        @par Postconditions
-        @code
-        this->host_type() == host_type::ipv6 && this->ipv6_address() == addr
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.2">
-            3.2.2. Host (rfc3986)</a>
-
-        @param addr The address to set.
-    */
-#ifdef BOOST_URL_DOCS
-    url& set_host( ipv6_address const& addr );
-#else
-    BOOST_URL_DECL
-    url&
-    set_host(
-        urls::ipv6_address const& addr);
-#endif
-
-    /** Set the host
-
-        The host is set to the specified plain
-        string, subject to the following rules:
-
-        @li If the string is a valid IPv4 address,
-        the address is parsed and the host is set
-        as if an instance of the equivalent
-        @ref urls::ipv4_address were passed instead.
-        In this case @ref url::host_type will return
-        @ref host_type::ipv4. Otherwise,
-
-        @li The plain string is percent-encoded and
-        the result is set as the reg-name for the
-        host. In this case @ref url::host_type will
-        return @ref host_type::name.
-
-        In all cases, if an authority did not
-        previously exist it is added by prepending
-        a double slash ("//") at the beginning of
-        the URL or after the scheme if a scheme is
-        present.
-
-        @par Postconditions
-        @code
-        this->encoded_host() == s
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.2">
-            3.2.2. Host (rfc3986)</a>
-
-        @param s The string to set.
-    */
-    BOOST_URL_DECL
-    url&
-    set_host(
-        string_view s);
-
-    /// @copydoc set_host(string_view)
-    BOOST_URL_DECL
-    url&
-    set_host(
-        pct_encoded_view s);
-
-    /** Set the host
-
-        The host is set to the specified percent-
-        encoded string, subject to the following
-        rules:
-
-        @li If the string is a valid IPv4 address,
-        the address is parsed and the host is set
-        as if an instance of the equivalent
-        @ref urls::ipv4_address were passed instead. In
-        this case @ref url::host_type will return
-        @ref host_type::ipv4. Or,
-
-        @li If the string is a valid IPv6 address
-        enclosed in square brackets ('[' and ']'),
-        the address is parsed and the host is set
-        as if an instance of the equivalent
-        @ref urls::ipv6_address were passed instead. In
-        this case @ref url::host_type will return
-        @ref host_type::ipv4. Or,
-
-        @li If the string is a valid IPvFuture address
-        enclosed in square brackets ('[' and ']'),
-        the address is parsed and the host is set
-        to the specified string. In this case
-        @ref url::host_type will return
-        @ref host_type::ipvfuture. Or,
-
-        @li If the string is a valid percent-encoded
-        string with no characters from the reserved
-        character set, then it is set as the encoded
-        host name. In this case @ref url::host_type
-        will return @ref host_type::name. Otherwise,
-
-        @li If the string does not contain a valid
-        percent-encoding for the host field, an
-        exception is thrown.
-
-        In all cases, if an authority did not
-        previously exist it is added by prepending
-        a double slash ("//") at the beginning of
-        the URL or after the scheme if a scheme is
-        present.
-        
-        @par Postconditions
-        @code
-        this->encoded_host() == s
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.2">
-            3.2.2. Host (rfc3986)</a>
-
-        @param s The string to set.
-
-        @throw std::exception the percent-encoding is invalid
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_host(string_view s);
-
-private:
-    char* set_port_impl(std::size_t n);
-public:
-
-    /** Remove the port
-
-        If a port is present, it is removed.
-        The remainder of the authority component
-        is left unchanged including the leading
-        double slash ("//").
-
-        @par Postconditions
-        @code
-        this->has_port() == false && this->port_number() == 0 && this->port() == ""
-        @endcode
-
-        @par Exception Safety
-        Throws nothing.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.3">
-            3.2.3. Port (rfc3986)</a>
-    */
-    BOOST_URL_DECL
-    url&
-    remove_port() noexcept;
-
-    /** Set the port
-
-        The port of the URL is set to the specified
-        integer, replacing any previous port.
-        If an authority did not
-        previously exist it is added by prepending
-        a double slash ("//") at the beginning of
-        the URL or after the scheme if a scheme is
-        present.
-
-        @par Postconditions
-        @code
-        this->has_port() == true && this->port_number() == n && this->port() == std::to_string(n)
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.3">
-            3.2.3. Port (rfc3986)</a>
-
-        @param n The port number to set.
-    */
-    BOOST_URL_DECL
-    url&
-    set_port(std::uint16_t n);
-
-    /** Set the port
-
-        The port of the URL is set to the specified
-        string, replacing any previous port. The string
-        must meet the syntactic requirements for PORT,
-        which consists only of digits. The string
-        may be empty. In this case the port is still
-        defined, however it is the empty string. To
-        remove the port call @ref remove_port.
-        If an authority did not
-        previously exist it is added by prepending
-        a double slash ("//") at the beginning of
-        the URL or after the scheme if a scheme is
-        present.
-
-        @par Postconditions
-        @code
-        this->has_port() == true && this->port_number() == n && this->port() == std::to_string(n)
-        @endcode
-
-        @par BNF
-        @code
-        port          = *DIGIT
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2.3">
-            3.2.3. Port (rfc3986)</a>
-
-        @param s The port string to set.
-
-        @throw std::exception `s` is not a valid port string.
-    */
-    BOOST_URL_DECL
-    url&
-    set_port(string_view s);
-
-    //--------------------------------------------
-
-    /** Remove the authority
-
-        The full authority component is removed
-        if present, which includes the leading
-        double slashes ("//"), the userinfo,
-        the host, and the port.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2">
-            3.2. Authority (rfc3986)</a>
-    */
-    BOOST_URL_DECL
-    url&
-    remove_authority() noexcept;
-
-    /** Set the authority
-
-        This function sets the authority component
-        to the specified encoded string. If a
-        component was present it is replaced.
-        Otherwise, the authority is added
-        including leading double slashes ("//").
-
-        The encoded string must be a valid
-        authority or else an exception is thrown.
-
-        @par BNF
-        @code
-        authority     = [ userinfo "@" ] host [ ":" port ]
-
-        userinfo      = *( unreserved / pct-encoded / sub-delims / ":" )
-        host          = IP-literal / IPv4address / reg-name
-        port          = *DIGIT
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.2">
-            3.2. Authority (rfc3986)</a>
-
-        @param s The authority string to set.
-
-        @throw std::exception `s` is not a valid authority.
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_authority(
-        string_view s);
-
-    //--------------------------------------------
-
-    /** Remove the origin component
-
-        The origin consists of the everything from the
-        beginning of the URL up to but not including
-        the path.
-
-        @par Exception Safety
-        Throws nothing.
-    */
-    BOOST_URL_DECL
-    url&
-    remove_origin() noexcept;
-
-    //--------------------------------------------
-    //
-    // Path
-    //
-    //--------------------------------------------
-
-private:
-    pos_t
-    segment(
-        std::size_t i) const noexcept;
-
-    char*
-    edit_segments(
-        std::size_t i0,
-        std::size_t i1,
-        std::size_t n,
-        std::size_t nseg);
-
-    BOOST_URL_DECL
-    void
-    edit_segments(
-        std::size_t i0,
-        std::size_t i1,
-        detail::any_path_iter&& it0,
-        detail::any_path_iter&& it1,
-        int abs_hint = -1);
-public:
-
-    /** Set whether the path is absolute.
-
-        This modifies the path as needed to
-        make it absolute or relative.
-
-        @return true on success.
-    */
-    BOOST_URL_DECL
-    bool
-    set_path_absolute(bool absolute);
-
-    /** Set the path.
-
-        This function validates the given percent-encoded
-        path according to the allowed grammar based
-        on the existing contents of the URL. If the
-        path is valid, the old path is replaced with
-        the new path. Otherwise, an exception is
-        thrown. The requirements for `s` are thus:
-
-        @li If `s` matches <em>path-empty</em>, that is
-        if `s.empty() == true`, the path is valid. Else,
-
-        @li If an authority is present (@ref has_authority
-        returns `true`), the path syntax must match
-        <em>path-abempty</em>. Else, if there is no
-        authority then:
-
-        @li If the new path starts with a forward
-        slash ('/'), the path syntax must match
-        <em>path-absolute</em>. Else, if the
-        path is rootless (does not start with '/'),
-        then:
-
-        @li If a scheme is present, the path syntax
-        must match <em>path-rootless</em>, otherwise
-
-        @li The path syntax must match <em>path-noscheme</em>.
-
-        @par BNF
-        @code
-        path-abempty  = *( "/" segment )
-        path-absolute = "/" [ segment-nz *( "/" segment ) ]
-        path-noscheme = segment-nz-nc *( "/" segment )
-        path-rootless = segment-nz *( "/" segment )
-        path-empty    = 0<pchar>
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set.
-
-        @throws std::invalid_argument invalid path.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.3"
-            >3.3. Path (rfc3986)</a>
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_path(
-        string_view s);
-
-    /** Set the path.
-
-        Sets the path of the URL to the specified
-        plain string. Any reserved characters in the
-        string will be automatically percent-encoded.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set. This string may
-        contain any characters, including nulls.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.3"
-            >3.3. Path (rfc3986)</a>
-
-        @see
-            @ref encoded_path,
-            @ref encoded_segments,
-            @ref is_path_absolute,
-            @ref normalize_path,
-            @ref set_encoded_path,
-            @ref set_path_absolute.
-    */
-    BOOST_URL_DECL
-    url&
-    set_path(
-        string_view s);
-
-    /// @copydoc set_path()
-    BOOST_URL_DECL
-    url&
-    set_path(
-        pct_encoded_view s);
-
-    /** Return the path segments
-
-        This function returns the path segments as
-        a bidirectional range.
-
-        @par BNF
-        @code
-        path          = [ "/" ] segment *( "/" segment )
-        @endcode
-
-        @par Exception Safety
-        Throws nothing.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.3"
-            >3.3. Path (rfc3986)</a>
-    */
-    BOOST_URL_DECL
-    segments_encoded
-    encoded_segments() noexcept;
-
-    /** Return the path segments
-
-        This function returns the path segments as
-        a bidirectional range.
-
-        @par BNF
-        @code
-        path          = [ "/" ] segment *( "/" segment )
-        @endcode
-
-        @par Exception Safety
-        Throws nothing.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-3.3"
-            >3.3. Path (rfc3986)</a>
-    */
-    urls::segments
-    segments() noexcept
-    {
-        return {*this};
-    }
-
-    //--------------------------------------------
-    //
-    // Query
-    //
-    //--------------------------------------------
-
-private:
-    struct raw_param
-    {
-        std::size_t pos;
-        std::size_t nk;
-        std::size_t nv;
-    };
-
-    raw_param
-    param(
-        std::size_t i) const noexcept;
-
-    char*
-    edit_params(
-        std::size_t i0,
-        std::size_t i1,
-        std::size_t n,
-        std::size_t nparam);
-
-    BOOST_URL_DECL
-    void
-    edit_params(
-        std::size_t i0,
-        std::size_t i1,
-        detail::any_query_iter&& it0,
-        detail::any_query_iter&& it1,
-        bool set_hint = false);
-public:
-
-    /** Remove the query.
-
-        If a query is present (@ref has_query
-        returns `true`), then the query is
-        removed including the leading `?`.
-
-        @par Exception Safety
-        Throws nothing.
-
-        @see
-            @ref has_query,
-            @ref set_encoded_query,
-            @ref set_query.
-    */
-    BOOST_URL_DECL
-    url&
-    remove_query() noexcept;
-
-    /** Set the query.
-
-        Sets the query of the URL to the specified
-        encoded string. The string must contain a
-        valid percent-encoding or else an exception
-        is thrown. When this function returns,
-        @ref has_query will return `true`.
-
-        @par BNF
-        @code
-        query           = *( pchar / "/" / "?" )
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set.
-
-        @throws std::invalid_argument bad encoding.
-
-        @see
-            @ref has_query,
-            @ref remove_query,
-            @ref set_query.
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_query(
-        string_view s);
-
-    /** Set the query.
-
-        Sets the query of the URL to the specified
-        plain string. Any reserved characters in the
-        string will be automatically percent-encoded.
-        When this function returns, @ref has_query
-        will return `true`.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set. This string may
-        contain any characters, including nulls.
-
-        @see
-            @ref has_query,
-            @ref remove_query,
-            @ref set_encoded_query.
-    */
-    BOOST_URL_DECL
-    url&
-    set_query(
-        string_view s);
-
-    /// @copydoc set_query()
-    BOOST_URL_DECL
-    url&
-    set_query(
-        pct_encoded_view s);
-
-    /** Return the query parameters
-
-        This function returns the query
-        parameters as a forward range of
-        key/value pairs.
-
-        Each string returned by the container
-        is percent-encoded.
-
-        @par BNF
-        @code
-        query-params    = [ query-param ] *( "&" [ query-param ] )
-
-        query-param     = key [ "=" value ]
-
-        @endcode
-    */
-    urls::params_encoded
-    encoded_params() noexcept
-    {
-        return urls::params_encoded(*this);
-    }
-
-    /** Return the query parameters
-
-        This function returns the query
-        parameters as a forward range of
-        key/value pairs where each returned
-        string has percent-decoding applied.
-
-        @par BNF
-        @code
-        query-params    = [ query-param ] *( "&" [ query-param ] )
-
-        query-param     = key [ "=" value ]
-        @endcode
-
-    */
-    urls::params
-    params()
-    {
-        return {*this};
-    }
-
-    //--------------------------------------------
-    //
-    // Fragment
-    //
-    //--------------------------------------------
-
-private:
-    char* set_fragment_impl(std::size_t n);
-public:
-
-    /** Remove the fragment.
-
-        If a fragment is present (@ref has_fragment
-        returns `true`), then the fragment is
-        removed including the leading `#`.
-
-        @par Exception Safety
-        Throws nothing.
-
-        @see
-            @ref has_fragment,
-            @ref set_encoded_fragment,
-            @ref set_fragment.
-    */
-    BOOST_URL_DECL
-    url&
-    remove_fragment() noexcept;
-
-    /** Set the fragment.
-
-        Sets the fragment of the URL to the specified
-        encoded string. The string must contain a
-        valid percent-encoding or else an exception
-        is thrown. When this function returns,
-        @ref has_fragment will return `true`.
-
-        @par BNF
-        @code
-        fragment      = *( pchar / "/" / "?" )
-        @endcode
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set.
-
-        @throws std::exception bad encoding.
-
-        @see
-            @ref has_fragment,
-            @ref remove_fragment,
-            @ref set_fragment.
-    */
-    BOOST_URL_DECL
-    url&
-    set_encoded_fragment(
-        string_view s);
-
-    /** Set the fragment.
-
-        Sets the fragment of the URL to the specified
-        plain string. Any reserved characters in the
-        string will be automatically percent-encoded.
-        When this function returns, @ref has_fragment
-        will return `true`.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @param s The string to set. This string may
-        contain any characters, including nulls.
-
-        @see
-            @ref has_fragment,
-            @ref remove_fragment,
-            @ref set_encoded_fragment.
-    */
-    BOOST_URL_DECL
-    url&
-    set_fragment(
-        string_view s);
-
-    /// @copydoc set_fragment()
-    BOOST_URL_DECL
-    url&
-    set_fragment(
-        pct_encoded_view s);
-
-    //--------------------------------------------
-    //
-    // Normalization
-    //
-    //--------------------------------------------
-private:
-    void
-    normalize_octets_impl(
-        int id,
-        grammar::lut_chars const& cs) noexcept;
-
-    void
-    decoded_to_lower_impl(int id) noexcept;
-
-    void
-    to_lower_impl(int id) noexcept;
-public:
-
-    /** Normalize the URL components
-
-        Applies Syntax-based normalization to
-        all components of the URL.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-6.2.2"
-            >6.2.2 Syntax-Based Normalization (rfc3986)</a>
-
-    */
-    BOOST_URL_DECL
-    url&
-    normalize();
-
-    /** Normalize the URL scheme
-
-        Applies Syntax-based normalization to the
-        URL scheme.
-
-        The scheme is normalized to lowercase.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-6.2.2"
-            >6.2.2 Syntax-Based Normalization (rfc3986)</a>
-
-    */
-    BOOST_URL_DECL
-    url&
-    normalize_scheme();
-
-    /** Normalize the URL authority
-
-        Applies Syntax-based normalization to the
-        URL authority.
-
-        Percent-encoding triplets are normalized
-        to uppercase letters. Percent-encoded
-        octets that correspond to unreserved
-        characters are decoded.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-6.2.2"
-            >6.2.2 Syntax-Based Normalization (rfc3986)</a>
-
-    */
-    BOOST_URL_DECL
-    url&
-    normalize_authority();
-
-    /** Normalize the URL path
-
-        Applies Syntax-based normalization to the
-        URL path.
-
-        Percent-encoding triplets are normalized
-        to uppercase letters. Percent-encoded
-        octets that correspond to unreserved
-        characters are decoded. Redundant
-        path-segments are removed.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-6.2.2"
-            >6.2.2 Syntax-Based Normalization (rfc3986)</a>
-
-    */
-    BOOST_URL_DECL
-    url&
-    normalize_path();
-
-    /** Normalize the URL query
-
-        Applies Syntax-based normalization to the
-        URL query.
-
-        Percent-encoding triplets are normalized
-        to uppercase letters. Percent-encoded
-        octets that correspond to unreserved
-        characters are decoded.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-6.2.2"
-            >6.2.2 Syntax-Based Normalization (rfc3986)</a>
-
-    */
-    BOOST_URL_DECL
-    url&
-    normalize_query();
-
-    /** Normalize the URL fragment
-
-        Applies Syntax-based normalization to the
-        URL fragment.
-
-        Percent-encoding triplets are normalized
-        to uppercase letters. Percent-encoded
-        octets that correspond to unreserved
-        characters are decoded.
-
-        @par Exception Safety
-        Strong guarantee.
-        Calls to allocate may throw.
-
-        @par Specification
-        @li <a href="https://datatracker.ietf.org/doc/html/rfc3986#section-6.2.2"
-            >6.2.2 Syntax-Based Normalization (rfc3986)</a>
-
-    */
-    BOOST_URL_DECL
-    url&
-    normalize_fragment();
-
-    //--------------------------------------------
-    //
-    // Resolution
-    //
-    //--------------------------------------------
-
-    friend
-    void
-    resolve(
-        url_view const& base,
-        url_view const& ref,
-        url& dest,
-        error_code& ec);
-
-    // hidden friend
-    friend
-    std::ostream&
-    operator<<(std::ostream& os, url const& u)
-    {
-        return os << u.string();
+        return (*this)=static_cast<
+            url_view_base const&>(u);
     }
 
 private:
-    //--------------------------------------------
-    //
-    // implementation
-    //
-    //--------------------------------------------
+    char* allocate(std::size_t);
+    void deallocate(char* s);
 
-    friend class static_url_base;
-
-    void check_invariants() const noexcept;
-
-    void
-    ensure_space(
-        std::size_t nchar,
-        std::size_t nseg,
-        std::size_t nparam);
-
-    char*
-    resize_impl(
-        int id,
-        std::size_t new_size);
-
-    char*
-    resize_impl(
-        int first,
-        int last,
-        std::size_t new_size);
-
-    char*
-    shrink_impl(
-        int id,
-        std::size_t new_size);
-
-    char*
-    shrink_impl(
-        int first,
-        int last,
-        std::size_t new_size);
-
-    BOOST_URL_DECL
-    bool
-    resolve(
-        url_view const& base,
-        url_view const& ref,
-        error_code& ec);
+    BOOST_URL_DECL void clear_impl() noexcept override;
+    BOOST_URL_DECL void reserve_impl(std::size_t) override;
 };
 
 //----------------------------------------------------------
@@ -1705,38 +272,15 @@ private:
 inline
 void
 resolve(
-    url_view const& base,
-    url_view const& ref,
-    url& dest,
+    url_view_base const& base,
+    url_view_base const& ref,
+    url_base& dest,
     error_code& ec)
 {
     BOOST_ASSERT(&dest != &base);
     BOOST_ASSERT(&dest != &ref);
-    dest.resolve(base, ref, ec);
+    dest.resolve_impl(base, ref, ec);
 }
-
-//----------------------------------------------------------
-
-/** Format the encoded URL to the output stream
-
-    This function serializes the encoded URL
-    to the output stream.
-
-    @par Example
-    @code
-    url u( "http://www.example.com/index.htm" );
-
-    std::cout << u << std::endl;
-    @endcode
-
-    @return A reference to the output stream, for chaining
-
-    @param os The output stream to write to.
-
-    @param u The URL to write.
-*/
-std::ostream&
-operator<<(std::ostream& os, url const& u);
 
 } // urls
 } // boost
@@ -1762,8 +306,7 @@ struct hash< ::boost::urls::url >
     std::size_t
     operator()(::boost::urls::url const& u) const noexcept
     {
-        std::hash<::boost::urls::url_view> h(salt_);
-        return h(static_cast<::boost::urls::url_view>(u));
+        return u.digest(salt_);
     }
 
 private:
@@ -1771,8 +314,6 @@ private:
 };
 } // std
 #endif
-
-
 
 #include <boost/url/impl/params.hpp>
 #include <boost/url/impl/params_encoded.hpp>
