+++ test/unit/Jamfile
@@ -48,7 +48,9 @@ local SOURCES =
     static_url.cpp
     string_view.cpp
     url.cpp
+    url_base.cpp
     url_view.cpp
+    url_view_base.cpp
     urls.cpp
     variant.cpp
     grammar/alnum_chars.cpp
