+++ test/unit/params.cpp
@@ -12,6 +12,7 @@
 #include <boost/url/params.hpp>
 
 #include <boost/url/url.hpp>
+#include <boost/url/url_view.hpp>
 #include <boost/core/ignore_unused.hpp>
 #include "test_suite.hpp"
