+++ test/unit/params_encoded.cpp
@@ -12,6 +12,7 @@
 #include <boost/url/params_encoded.hpp>
 
 #include <boost/url/url.hpp>
+#include <boost/url/url_view.hpp>
 #include "test_suite.hpp"
 
 namespace boost {
