+++ test/unit/segments.cpp
@@ -12,6 +12,7 @@
 #include <boost/url/segments.hpp>
 
 #include <boost/url/url.hpp>
+#include <boost/url/url_view.hpp>
 #include <boost/static_assert.hpp>
 #include <boost/core/ignore_unused.hpp>
 #include <initializer_list>
