+++ test/unit/segments_encoded.cpp
@@ -12,6 +12,7 @@
 #include <boost/url/segments_encoded.hpp>
 
 #include <boost/url/url.hpp>
+#include <boost/url/url_view.hpp>
 #include <boost/static_assert.hpp>
 #include <initializer_list>
 #include <iterator>
